Fix OAuth login not creating browser session
The OAuth login flow was storing "user_id" directly in the session, but
the authentication plug looks for "session_token". This meant users were
redirected to their profile page after GitHub OAuth but never actually
logged in.

Now the OAuth controller properly creates browser sessions using
LoginController.start_session/3 for non-TFA users and pre-creates
session tokens for TFA users.
